move actual metrics computation to @lionweb/utilities
