+++ packages/cli/src/serialization/measurer.ts
@@ -1,6 +1,6 @@
 import {extname} from "path"
-import {conceptsOf, groupBy, Language, MemoisingSymbolTable, MetaPointer, SerializationChunk} from "@lionweb/core"
-import {readChunk, tryLoadAllAsLanguages, writeJsonAsFile} from "@lionweb/utilities"
+import {Language} from "@lionweb/core"
+import {measure, readChunk, tryLoadAllAsLanguages, writeJsonAsFile} from "@lionweb/utilities"
 import {separate} from "../language-aware-args.js"
 
 export const executeMeasureCommand = async (args: string[]) => {
@@ -17,51 +17,5 @@ const measureSerializationChunk = async (path: string, languages: Language[]) =>
     const metricsPath = extLessPath + ".metrics.json"
     writeJsonAsFile(metricsPath, measure(chunk, languages))
     console.log(`Wrote metrics for ${path} --> ${metricsPath}`)
-}
-
-// TODO  move all of this to @lionweb/utilities (in the end)
-
-type ClassifierInstantiationMetric = {
-    name?: string
-    key: string // == key of classifier
-    language: string // == key of language
-    version: string
-    count: number
-    // TODO  add property to say the classifier is a concept, or annotation (or enum)?
-}
-
-type Metrics = {
-    instantiations: ClassifierInstantiationMetric[]
-    unusedConcreteConcepts: MetaPointer[]
-}
-
-const measure = (serializationChunk: SerializationChunk, languages: Language[]): Metrics => {
-    const symbolTable = new MemoisingSymbolTable(languages)
-
-    const metaPointerAsText = (classifier: MetaPointer) => `${classifier.key}:${classifier.language}:${classifier.version}`
-
-    // group nodes by classifier key, language, and version:
-    const instantiationsPerId = groupBy(serializationChunk.nodes, ({ classifier }) => metaPointerAsText(classifier))
-
-    // map grouped nodes to instantiations with count:
-    const instantiations = Object.values(instantiationsPerId).map((nodes) => ({
-        ...nodes[0].classifier,
-        name: symbolTable.entityMatching(nodes[0].classifier)?.name,
-        count: nodes.length
-    }))
-
-    // compute all concrete concepts for the language:
-    const concreteConcepts = languages.flatMap(conceptsOf).filter((concept) => !concept.abstract)
-    const unusedConcreteConcepts = concreteConcepts.filter((concept) => !(metaPointerAsText(concept.metaPointer()) in instantiationsPerId))
-
-    // return the metrics object:
-    return {
-        instantiations,
-        unusedConcreteConcepts: unusedConcreteConcepts.map(concept => ({
-            language: concept.language.key,
-            version: concept.language.version,
-            key: concept.key
-        }))
-    }
 }
