+++ packages/utilities/src/serialization/index.ts
@@ -1,4 +1,5 @@
 export {looksLikeSerializedLanguages, readChunk, tryLoadAllAsLanguages, tryLoadAsLanguages} from "./chunk.js"
+export {measure} from "./measurer.js"
 export {orderedSerializationChunk} from "./ordering.js"
 export {sortedSerializationChunk} from "./sorting.js"
 export {genericAsTreeText} from "./textualizer.js"
